componentsにImagesChangeとMainContainerに分けて配置
